Create index.html
This is about Power Corp Group, Our group is out to proffer solutions to the persistent issue of epileptic power problem  we experience in this part of the world.  

Our target audience are both medium and small size entrepreneurs and all citizens in need of power, to power their daily lifes.

Our UVP is to be a Power enterprise that will ensure availability of power, efficiency in power supply and profitability to both power providers and consumers..  

The impact our solution will create will eliminate the problems we currently face in all facet of the Nigeria economy.